letstry/requirements.txt
openai>=1.0.0
Flask>=2.0.0
numpy>=1.25.0
python-dotenv>=1.0.0
